i32640575 deletion compatible with "branching": a reebill can only be deleted if its predecessor has no unissued version

+++ reebill/ui/billedit.js
@@ -580,8 +580,16 @@ function reeBillReady() {
             var sequences = selectedRecords.map(function(rec) {
                 return rec.data.sequence;
             });
-            deleteReebills(sequences);
-            reeBillStore.save();
+
+            Ext.Msg.confirm('Confirm deletion',
+                'Are you sure you want to delete the latest version of reebill '
+                + selected_account + '-' + sequences + '?', function(answer) {
+                    if (answer == 'yes') {
+                        deleteReebills(sequences);
+                    }
+            });
+
+            reeBillStore.reload();
         }
     })
